+++ src/types.ts
@@ -28,3 +28,13 @@ export interface ILoggedInContextType {
 export interface ChildrenProps {
   children: any;
 }
+
+export type LoginProps = {
+  newMember: (value: boolean) => void;
+};
+
+export interface modalProps {
+  openModal: boolean;
+  userName: string;
+  setOpenModal: (value: boolean) => void;
+}
